Add a Job class
Use of the Pheanstalk Job class in ConsumerInterface makes this
library a leaky abstraction with respect to its internal use of
Pheanstalk, which is an implementation detail that calling code
shouldn't need to be aware of.

This change creates a Job class with an equivalent API that can
wrap the data from a Pheanstalk Job instance. This removes the
leaky abstraction, making it possible to change the underlying
library in the future if need be.